--- frontend/src/components/Dashboard/Dashboard.tsx.preview
@@ -1,1 +1,0 @@
-<DashboardContent />
